epoch-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai>=1.0.0
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
